webapp-di6tpp/requirement_packages.txt
proxmoxer
requests
mysql-connector-python
jinja2
itsdangerous
uvicorn
httpx
starlette
python-multipart
fastapi
toml
SQLAlchemy
sqlalchemy-utils
pymysql
pycryptodome
passlib[argon2]
pydantic
alembic
python-dotenv
psutil
redis 
paramiko
PyJWT
pytest==6.2.5
httpx==0.28.1
pytest-anyio
asgi-lifespan